# packages/api/openapi.yaml
openapi: 3.0.3
info:
  title: FleetSync API
  description: |
    FleetSync is a comprehensive fleet management system with real-time tracking,
    telemetry ingestion, warehouse management, and driver allocation.
  version: 1.0.0
  contact:
    name: FleetSync Team
    email: api@fleetsync.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:4000
    description: Development server
  - url: https://api.fleetsync.com
    description: Production server
...
paths:
...
  /vehicles/{id}:
    get:
      summary: Get vehicle by ID
      tags: [Vehicles]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Vehicle details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
  /telemetry:
    post:
      summary: Submit single telemetry event
      tags: [Telemetry]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TelemetryEvent'
      responses:
        '201':
          description: Telemetry received
  /telemetry/bulk:
    post:
      summary: Submit bulk telemetry events
      tags: [Telemetry]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                events:
                  type: array
                  items:
                    $ref: '#/components/schemas/TelemetryEvent'
      responses:
        '201':
          description: Bulk telemetry received
  /telemetry/vehicle/{vehicleId}:
    get:
      summary: Get telemetry for vehicle
      tags: [Telemetry]
      parameters:
        - name: vehicleId
          in: path
          required: true
          schema:
            type: string
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Vehicle telemetry data
          content:
            application/json:
              schema:
                type: object
                properties:
                  vehicleId:
                    type: string
                  count:
                    type: integer
                  events:
                    type: array
                    items:
                      $ref: '#/components/schemas/TelemetryEvent'
  /trips:
    get:
      summary: Get all trips
      tags: [Trips]
      responses:
        '200':
          description: List of trips
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Trip'
    post:
      summary: Create new trip
      tags: [Trips]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TripCreate'
      responses:
        '201':
          description: Trip created
  /trips/{id}/assign:
    post:
      summary: Assign trip to vehicle
      tags: [Trips]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                vehicleId:
                  type: string
      responses:
        '200':
          description: Trip assigned
...
components:
  schemas:
...
    Vehicle:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        lat:
          type: number
        lng:
          type: number
        status:
          type: string
          enum: [empty, loaded, maintenance]
        driver:
          type: string
        fuel_level:
          type: number
        driver_hours_today:
          type: number
        needs_maintenance:
          type: boolean
    TelemetryEvent:
      type: object
      required: [vehicleId, lat, lng, timestamp]
      properties:
        vehicleId:
          type: string
        lat:
          type: number
        lng:
          type: number
        speed:
          type: number
        heading:
          type: number
        fuel_level:
          type: number
        timestamp:
          type: string
          format: date-time
    Trip:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
        origin:
          $ref: '#/components/schemas/Location'
        destination:
          $ref: '#/components/schemas/Location'
        type:
          type: string
        passengers:
          type: integer
    TripCreate:
      type: object
      required: [origin, destination, type]
      properties:
        origin:
          $ref: '#/components/schemas/Location'
        destination:
          $ref: '#/components/schemas/Location'
        type:
          type: string
        passengers:
          type: integer
    Location:
      type: object
      properties:
        lat:
          type: number
        lng:
          type: number
        address:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []